Refactor: move tests for tokenize() to its own module
